modularity
taking thing that can be reused and making them separate widgets and just calling them
